Refactor to fix wrong imports
