+++ src/places/pages/UserPlaces.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useParams } from "react-router-dom";
 
-import PlaceList from "../components/PlaceList";
+import PlaceList from "../components/PlaceList/PlaceList";
 
 const DUMMY_PLACES = [
   {
